[QoS] total time spent talking to Cassandra (#2687)
* total-time

* [QoS] Client config (#2690)

* qos config

* respect max backoff itme

* [QoS] [Refactor] Query Weights (#2697)

* query weights

* extra tests

* [QoS] Number of rows per query (#2698)

* num rows

* checkstyle

* fix tests

* no int casting

* fix numRows calculation on batch_mutate

+++ qos-service-impl/src/test/java/com/palantir/atlasdb/qos/QosServiceTest.java
@@ -39,20 +39,20 @@ public class QosServiceTest {
     public void defaultsToNoLimit() {
         when(config.get()).thenReturn(configWithLimits(ImmutableMap.of()));
 
-        assertThat(resource.getLimit("foo")).isEqualTo(Integer.MAX_VALUE);
+        assertThat(resource.getLimit("foo")).isEqualTo(Long.MAX_VALUE);
     }
 
     @Test
     public void canLiveReloadLimits() {
         when(config.get())
-                .thenReturn(configWithLimits(ImmutableMap.of("foo", 10)))
-                .thenReturn(configWithLimits(ImmutableMap.of("foo", 20)));
+                .thenReturn(configWithLimits(ImmutableMap.of("foo", 10L)))
+                .thenReturn(configWithLimits(ImmutableMap.of("foo", 20L)));
 
-        assertEquals(10, resource.getLimit("foo"));
-        assertEquals(20, resource.getLimit("foo"));
+        assertEquals(10L, resource.getLimit("foo"));
+        assertEquals(20L, resource.getLimit("foo"));
     }
 
-    private QosServiceRuntimeConfig configWithLimits(Map<String, Integer> limits) {
+    private QosServiceRuntimeConfig configWithLimits(Map<String, Long> limits) {
         return ImmutableQosServiceRuntimeConfig.builder().clientLimits(limits).build();
     }
 }
